🚨 Extract `folders` from `tree_db.root_paths`
TG-2031
