+++ src/youwol/backends/tree_db/root_paths.py
@@ -3,8 +3,6 @@ import asyncio
 import itertools
 import uuid
 
-from collections.abc import Coroutine
-
 # third parties
 from fastapi import APIRouter, Depends, HTTPException
 from starlette.requests import Request
@@ -26,36 +24,31 @@ from youwol.utils import (
 )
 from youwol.utils.context import Context
 from youwol.utils.http_clients.tree_db_backend import (
-    ChildrenResponse,
     DefaultDriveResponse,
     DriveBody,
     DriveResponse,
     DrivesResponse,
     FolderBody,
-    FolderResponse,
     Group,
     GroupsResponse,
-    ItemResponse,
-    PathResponse,
     PurgeResponse,
     RenameBody,
 )
 
 # relative
-from .routers import router_entities, router_items
+from .routers import router_entities, router_folders, router_items
+from .routers.folders import purge_folder
 from .utils import (
+    create_folder,
     db_delete,
     db_get,
     db_post,
     db_query,
     doc_to_drive_response,
-    doc_to_folder,
     entities_children,
     folder_to_doc,
     get_drive,
     get_folder,
-    get_folders_rec,
-    get_parent,
     item_to_doc,
     list_deleted,
 )
@@ -63,6 +56,7 @@ from .utils import (
 router = APIRouter(tags=["treedb-backend"])
 router.include_router(router_entities)
 router.include_router(router_items)
+router.include_router(router_folders)
 flatten = itertools.chain.from_iterable
 
 
@@ -302,7 +296,7 @@ async def _ensure_folder(
             if e_folder.status_code != 404:
                 raise e_folder
             await ctx.warning("Folder does not exist yet, start creation")
-            return await _create_folder(
+            return await create_folder(
                 parent_folder_id=parent_folder_id,
                 folder=FolderBody(name=name, folderId=folder_id),
                 configuration=configuration,
@@ -418,270 +412,6 @@ async def get_default_drive(
         Description of the drive.
     """
     return await _get_default_drive(request, group_id, configuration)
-
-
-async def _create_folder(
-    parent_folder_id: str,
-    folder: FolderBody,
-    configuration: Configuration,
-    context: Context,
-):
-    async with context.start(action="_create_folder") as ctx:  # type: Context
-        folders_db, _ = (
-            configuration.doc_dbs.folders_db,
-            configuration.doc_dbs.drives_db,
-        )
-        parent = await get_parent(
-            parent_id=parent_folder_id, configuration=configuration, context=ctx
-        )
-
-        doc = {
-            "folder_id": folder.folderId or str(uuid.uuid4()),
-            "name": folder.name,
-            "parent_folder_id": parent_folder_id,
-            "group_id": parent["group_id"],
-            "type": folder.kind,
-            "metadata": folder.metadata,
-            "drive_id": parent["drive_id"],
-        }
-        await db_post(docdb=folders_db, doc=doc, context=ctx)
-
-        response = doc_to_folder(doc)
-        return response
-
-
-@router.put(
-    "/folders/{parent_folder_id}",
-    summary="create a folder",
-    response_model=FolderResponse,
-)
-async def create_folder(
-    request: Request,
-    parent_folder_id: str,
-    folder: FolderBody,
-    configuration: Configuration = Depends(get_configuration),
-) -> FolderResponse:
-    """
-    Creates a new folder.
-
-    Parameters:
-        request: Incoming request.
-        parent_folder_id: ID of the parent folder.
-        folder: folder properties.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the folder.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="create_folder",
-        body=folder,
-        with_attributes={"parent_folder_id": parent_folder_id},
-    ) as ctx:  # type: Context
-        response = await _create_folder(
-            parent_folder_id=parent_folder_id,
-            folder=folder,
-            configuration=configuration,
-            context=ctx,
-        )
-        return response
-
-
-@router.post(
-    "/folders/{folder_id}", summary="update a folder", response_model=FolderResponse
-)
-async def update_folder(
-    request: Request,
-    folder_id: str,
-    body: RenameBody,
-    configuration: Configuration = Depends(get_configuration),
-) -> FolderResponse:
-    """
-    Updates a folder properties.
-
-    Parameters:
-        request: Incoming request.
-        folder_id: ID of the folder.
-        body: Update details.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the drive.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="update_folder",
-        body=body,
-        with_attributes={"folder_id": folder_id},
-    ) as ctx:  # type: Context
-        folders_db = configuration.doc_dbs.folders_db
-        doc = await db_get(
-            partition_keys={"folder_id": folder_id}, docdb=folders_db, context=ctx
-        )
-        doc = {**doc, **{"name": body.name}}
-        await db_post(docdb=folders_db, doc=doc, context=ctx)
-
-        response = doc_to_folder(doc)
-        return response
-
-
-@router.get(
-    "/folders/{folder_id}",
-    summary="Retrieves properties of a folder.",
-    response_model=FolderResponse,
-)
-async def get_folder_details(
-    request: Request,
-    folder_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> FolderResponse:
-    """
-    Retrieves properties of a folder.
-
-    Parameters:
-        request: Incoming request.
-        folder_id: ID of the folder.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the folder.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="get_folder",
-        with_attributes={"folder_id": folder_id},
-    ) as ctx:  # type: Context
-        response = await get_folder(
-            folder_id=folder_id, configuration=configuration, context=ctx
-        )
-        return response
-
-
-@router.get(
-    "/folders/{folder_id}/path",
-    summary="Retrieves the full path of a folder.",
-    response_model=PathResponse,
-)
-async def get_path_folder(
-    request: Request,
-    folder_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> PathResponse:
-    """
-    Retrieves the full path of a folder.
-
-    Parameters:
-        request: Incoming request.
-        folder_id: ID of the folder.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the path.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="get_path_folder",
-        with_attributes={"folder_id": folder_id},
-    ) as ctx:  # type: Context
-        folder = await get_folder(
-            folder_id=folder_id, configuration=configuration, context=ctx
-        )
-        folders, drive = await get_folders_rec(
-            folder_id=folder_id,
-            drive_id=folder.driveId,
-            configuration=configuration,
-            context=ctx,
-        )
-
-        response = PathResponse(folders=folders, drive=drive)
-        return response
-
-
-@router.get(
-    "/folders/{folder_id}/children",
-    summary="Query the children of a folder or drive.",
-    response_model=ChildrenResponse,
-)
-async def children(
-    request: Request,
-    folder_id: str,
-    configuration: Configuration = Depends(get_configuration),
-) -> ChildrenResponse:
-    """
-    Query the children of a folder or drive.
-
-    Parameters:
-        request: Incoming request
-        folder_id: parent folder's ID (or parent's drive ID to request a drive's children).
-        configuration: Injected configuration of the service.
-
-    Return:
-        Description of the children.
-    """
-    async with Context.start_ep(
-        request=request, action="children"
-    ) as ctx:  # type: Context
-        response = await entities_children(
-            folder_id=folder_id, configuration=configuration, context=ctx
-        )
-        return response
-
-
-@router.delete("/folders/{folder_id}", summary="delete a folder and its content")
-async def queue_delete_folder(
-    request: Request,
-    folder_id: str,
-    configuration: Configuration = Depends(get_configuration),
-):
-    """
-    Queues a folder for deletion (moves into the 'trash').
-
-    Parameters:
-        request: Incoming request
-        folder_id: Folder's ID.
-        configuration: Injected configuration of the service.
-
-    Return:
-        Empty JSON response.
-    """
-    async with Context.start_ep(
-        request=request,
-        action="queue_delete_folder",
-        with_attributes={"itemId": folder_id},
-    ) as ctx:  # type: Context
-        dbs = configuration.doc_dbs
-        folders_db, _, deleted_db = (
-            dbs.folders_db,
-            dbs.drives_db,
-            dbs.deleted_db,
-        )
-
-        doc = await db_get(
-            partition_keys={"folder_id": folder_id}, docdb=folders_db, context=ctx
-        )
-
-        doc = {
-            "deleted_id": doc["folder_id"],
-            "drive_id": doc["drive_id"],
-            "type": doc["type"],
-            "kind": "folder",
-            "name": doc["name"],
-            "parent_folder_id": doc["parent_folder_id"],
-            "related_id": "",
-            "group_id": doc["group_id"],
-            "metadata": doc["metadata"],
-        }
-
-        deleted_db = configuration.doc_dbs.deleted_db
-
-        await db_post(doc=doc, docdb=deleted_db, context=ctx)
-        await db_delete(
-            docdb=folders_db,
-            doc={"folder_id": doc["deleted_id"], "group_id": doc["group_id"]},
-            context=ctx,
-        )
-        return {}
 
 
 @router.delete(
@@ -818,54 +548,3 @@ async def purge_drive(
             items=list_items + deleted.items,
         )
         return response
-
-
-async def purge_folder(
-    request: Request,
-    drive_id: str,
-    folder_id: str,
-    skip_folders: set[str],
-    skip_items: set[str],
-    configuration: Configuration,
-    context: Context,
-) -> tuple[list[Coroutine], list[Coroutine], list[ItemResponse]]:
-    async with context.start(action="purge folder") as ctx:
-        doc_dbs = configuration.doc_dbs
-        content = await entities_children(
-            folder_id=folder_id, configuration=configuration, context=ctx
-        )
-
-        delete_items = [
-            db_delete(docdb=doc_dbs.items_db, doc=item_to_doc(f), context=ctx)
-            for f in content.items
-            if f.itemId not in skip_items
-        ]
-        delete_folders = [
-            db_delete(docdb=doc_dbs.folders_db, doc=folder_to_doc(f), context=ctx)
-            for f in content.folders
-            if f.folderId not in skip_items
-        ]
-
-        skip_items = skip_items.union({f.itemId for f in content.items})
-
-        children_folder = [
-            purge_folder(
-                request=request,
-                drive_id=drive_id,
-                folder_id=f.folderId,
-                context=ctx,
-                skip_folders=skip_folders,
-                skip_items=skip_items,
-                configuration=configuration,
-            )
-            for f in content.folders
-            if f.folderId not in skip_folders
-        ]
-
-        deletion_rec_folders = await asyncio.gather(*children_folder)
-
-        return (
-            delete_items + list(flatten([d for d, _, _ in deletion_rec_folders])),
-            delete_folders + list(flatten([d for _, d, _ in deletion_rec_folders])),
-            content.items + list(flatten([d for _, _, d in deletion_rec_folders])),
-        )
