+++ src/youwol/backends/tree_db/routers/__init__.py
@@ -1,3 +1,4 @@
 # relative
 from .entities import router as router_entities
+from .folders import router as router_folders
 from .items import router as router_items
